+++ src/test/java/org/sonar/plugins/delphi/pmd/rules/EmptyBracketsRuleTest.java
@@ -23,7 +23,7 @@ public class EmptyBracketsRuleTest extends BasePmdRuleTest {
   }
 
   @Test
-  public void testMethodCallEmptyBracketsShouldAddIssue() {
+  public void testInvocationOfUnknownMethodShouldAddIssue() {
     DelphiTestUnitBuilder builder =
         new DelphiTestUnitBuilder()
             .appendImpl("procedure Test;")
@@ -36,5 +36,57 @@ public class EmptyBracketsRuleTest extends BasePmdRuleTest {
     assertIssues()
         .hasSize(1)
         .areExactly(1, ruleKeyAtLine("EmptyBracketsRule", builder.getOffset() + 3));
+  }
+
+  @Test
+  public void testInvocationOfKnownMethodShouldAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("procedure MyProcedure;")
+            .appendImpl("procedure Test;")
+            .appendImpl("begin")
+            .appendImpl("  MyProcedure();")
+            .appendImpl("end;");
+
+    execute(builder);
+
+    assertIssues()
+        .hasSize(1)
+        .areExactly(1, ruleKeyAtLine("EmptyBracketsRule", builder.getOffset() + 3));
+  }
+
+  @Test
+  public void testExplicitArrayConstructorShouldNotAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("type")
+            .appendDecl("  TIntArray = array of Integer;")
+            .appendImpl("procedure Test;")
+            .appendImpl("var")
+            .appendImpl("  Foo: TIntArray;")
+            .appendImpl("begin")
+            .appendImpl("  Foo := TIntArray.Create();")
+            .appendImpl("end;");
+
+    execute(builder);
+
+    assertIssues().isEmpty();
+  }
+
+  @Test
+  public void testInvocationOfProcVarShouldNotAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendDecl("procedure MyProcedure;")
+            .appendImpl("procedure Test;")
+            .appendImpl("var")
+            .appendImpl("  ProcVar: procedure;")
+            .appendImpl("begin")
+            .appendImpl("  ProcVar();")
+            .appendImpl("end;");
+
+    execute(builder);
+
+    assertIssues().isEmpty();
   }
 }